Split test and components for DocumentsList
